implemented changepoint in first half not ok

+++ project/pettitt/pettitt.cpp
@@ -24,7 +24,7 @@ std::tuple<double, double, int> Pettitt::test2(std::deque<double> const &x, std:
       return 2*rankPartialSum[x-1]-x*(n+1);
     };
     Eigen::ArrayXd Uk = k.unaryExpr(func);
-    size_t K;  // K will take into account 0 numbering, so this will be 9 with the one in R is 10
+    int K;  // K will take into account 0 numbering, so this will be 9 with the one in R is 10
     double U = Uk.abs().maxCoeff(&K);
     double pval = std::min(1.0,2*std::exp((-6*U*U)/(n*n*(n+1))));
     return std::make_tuple(pval, U, K);
